pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ha-agent"
version = "0.1.0"
description = "Home Assistant energy optimization AI agent using Claude"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "aiohttp>=3.9",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "jinja2>=3.1",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
ha-agent = "ha_agent.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/ha_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
